% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aligner.R
\name{align_spacer_seq}
\alias{align_spacer_seq}
\title{Align spacer sequence}
\usage{
align_spacer_seq(query_seq, grna_spacer, pam_pattern = "NGG")
}
\arguments{
\item{query_seq}{the target sequence (typically ~50-100 bp)}

\item{grna_spacer}{the 20-bp gRNA spacer sequence}

\item{pam_pattern}{the PAM (default: NGG)}
}
\value{
a data frame summarizing the alignment results
}
\description{
Align spacer sequence
}
