% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aligner.R
\name{align_spacer_to_genome}
\alias{align_spacer_to_genome}
\title{Align a spacer sequence to the genome}
\usage{
align_spacer_to_genome(
  grna_spacer,
  ref_genome = paste0(.get_config_path("REF_GENOME_DIR"), "hg38_main_chroms"),
  pam = NA
)
}
\arguments{
\item{grna_spacer}{gRNA spacer sequence (written in 5' -> 3' direction)}

\item{ref_genome}{file path to the reference genome}

\item{pam}{(optional) the PAM sequence}
}
\value{
a GAlignments object containing the alignmenet
}
\description{
Aligns a short (20pb) spacer sequence to the genome.
}
\examples{
alignment <- align_spacer_to_genome("ACTGATAGGGGTCGCGGTAG")
}
