% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicate_aware_guide_seq.R
\name{annotate_clustered_count_df_with_homology}
\alias{annotate_clustered_count_df_with_homology}
\title{Annotate clustered count df with homology}
\usage{
annotate_clustered_count_df_with_homology(clustered_count_df, homology_df)
}
\arguments{
\item{clustered_count_df}{output of \code{cluster_loci()}}

\item{homology_df}{output of \code{load_crispritz_output()}}
}
\description{
Add colummns:
\itemize{
\item \code{homology_has_hit} (indicating whether there is a homology hit inside the window)
\item \code{homology_n_mismatches} (indicating the number of mismatches between aligned spacer and protospacer sequence)
\item \code{homology_n_bulges} (indicating number of bulges between aligned spacer and protospacer sequence)
\item \code{homology_posit} (indicating the position of the start of the protospacer)
\item \code{homology_strand} (indicating whether the protospacer is on the plus or minus strand)
\item \code{homology_dna} (aligned protospacer sequence)
\item \code{homology_gRNA} (aligned spacer sequence)
}
}
\examples{
homology_df <- load_crispritz_output("/Users/timbarry/research_offsite/external/bauer-lab/guideseq_elane/crispritz/crispritz_CCCCGGCAGAAACGTCCGCG.hg38.targets.txt")
elane_dir <- paste0(.get_config_path("LOCAL_BAUER_LAB_DATA_DIR"), "guideseq_elane/")
clustered_count_df <- readRDS(paste0(elane_dir, "count_tables_no_multimap/combined_count_df.rds")) |>
dplyr::filter(treated & cell_type == "CD34" & cas9_variant == "wt_cas9" & replicate_id \%in\% 1:2 & chr != "chrM") |>
dplyr::select(chr, coord, strand, umi_count, primer_type, replicate_id) |>
cluster_loci()

annotated_clustered_count_df <- annotate_clustered_count_df_with_homology(clustered_count_df, homology_df)
}
