% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfd_score.R
\name{calulate_cfd_score}
\alias{calulate_cfd_score}
\title{Calculate CFD score}
\usage{
calulate_cfd_score(homology_dna, homology_gRNA, homology_has_hit)
}
\arguments{
\item{homology_dna}{character vector containing CRISPRitz-aligned DNA sequences.}

\item{homology_gRNA}{character vector containing CRISPRitz-aligned gRNA sequences.}

\item{homology_has_hit}{logical vector indicating whether each sequence pair has a CRISPRitz hit.}
}
\value{
numeric vector of CFD scores.
}
\description{
Calculate CFD score
}
\examples{
homology_df <- load_crispritz_output("/Users/timbarry/research_offsite/external/bauer-lab/guideseq_elane/crispritz_CCCCGGCAGAAACGTCCGCG.hg38.targets.txt")
n_run_df <- load_n_run_bed("/Users/timbarry/research_offsite/ref_genome_dir/hg38_N_runs_min10.bed")
elane_dir <- paste0(.get_config_path("LOCAL_BAUER_LAB_DATA_DIR"), "guideseq_elane/")
clustered_count_df <- readRDS(paste0(elane_dir, "count_tables_no_multimap/combined_count_df.rds")) |>
dplyr::filter(treated & cell_type == "CD34" & cas9_variant == "wt_cas9" & replicate_id \%in\% 1:2 & chr != "chrM") |>
dplyr::select(chr, coord, strand, umi_count, primer_type, replicate_id) |>
cluster_loci()
annotated_clustered_count_df <- annotate_clustered_count_df(clustered_count_df = clustered_count_df, homology_df = homology_df, n_run_df = n_run_df)
window_info <- annotated_clustered_count_df |> dplyr::arrange(-umi_count) |> dplyr::slice(4)
score <- calulate_cfd_score(homology_dna = window_info$homology_dna,
homology_gRNA = window_info$homology_gRNA,
homology_has_hit = window_info$homology_has_hit)
}
