% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_amplicon_seq.R
\name{collapse_count_matrix_by_allele}
\alias{collapse_count_matrix_by_allele}
\title{Collapse allele table}
\usage{
collapse_count_matrix_by_allele(
  count_matrix,
  allele_df,
  bucket_breaks = c(seq(0L, 49L), Inf),
  bucket_labels = c(seq(1L, 49L), "50+")
)
}
\arguments{
\item{count_matrix}{an allele count matrix}

\item{allele_df}{the allele data frame}

\item{bucket_breaks}{integer vector indicating the levels to break the mutation lengths into}

\item{bucket_labels}{character vector indicating the bucket labels}
}
\value{
the collapsed matrix
}
\description{
Collapse allele table
}
\examples{
data_list <- readRDS("/Users/timbarry/research_offsite/external/bauer-lab/rhampseq_bcl11a/multivariate_count_tables/combined/CRISPResso_on_1450_OT_0000.rds")
count_matrix <- data_list$count_matrix
allele_df <- data_list$allele_df
covariate_df <- data_list$covariate_df
collapsed_matrix_list <- collapse_count_matrix_by_allele(count_matrix, allele_df)
collapsed_res <- compute_basic_statistics_on_multivariate_amplicon_seq_data(collapsed_matrix_list$collapsed_matrix, covariate_df)

}
