% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_crispresso_output.R
\name{compute_base_editing_count_table}
\alias{compute_base_editing_count_table}
\title{Compute base editing count table}
\usage{
compute_base_editing_count_table(
  crispresso_dir,
  indel_positions = c(17L, 18L),
  bystander_substitution_positions = seq(3L, 11),
  intended_substitution_positions = integer(0L),
  from_nuc = "A",
  to_nuc = "G"
)
}
\arguments{
\item{crispresso_dir}{file path to a CRISPResso output directory}

\item{indel_positions}{bases over which to quantify indels; default bases flanking PAM}

\item{bystander_substitution_positions}{bases over which to quantify bystander substitutions; default positions 3-11 from distal end of protospacer}

\item{intended_substitution_positions}{bases over which to quantify intended substitutions; default integer(0L) (i.e., empty set)}

\item{from_nuc}{original nucleotide; default "A"}

\item{to_nuc}{converted nucleotide; default "G"}
}
\value{
a 1-row data frame with columns \code{n_reads}, \code{n_reads_w_intended_substitution}, \code{n_reads_w_bystander_substitution}, \code{n_reads_w_indel}, \code{n_reads_w_modification}
}
\description{
This function takes a CRISPResso directory as input. It outputs a data frame with one row containing the following columns:
\itemize{
\item n_reads: total number of (aligned) reads for this sample.
\item n_reads_w_intended_substitution: number of reads harboring the intended substitution.
\item n_reads_w_bystander_substitution: number of reads harboring a bystander substitution.
\item n_reads_w_indel: number of reads harboring an indel.
\item n_reads_w_modification: number of reads harboring an intended substitution OR bystander substitution OR indel.
}
}
\details{
Identify modification locations using 1-based indexing relative to the distal end of the protospacer sequence. For example, the default \code{indel_positions} is \code{c(17L, 18L)} (i.e., the bases immediately adjacent to the nicking site), while the default \code{bystander_substitution_positions} is \code{seq(3L, 11)} (i.e., positions 3-11 from the distal end of the protospacer).

Specify the substitution type using the \code{from_nuc} and \code{to_nuc} arguments; for an adenine base editor, \code{from_nuc = "A"} and \code{to_nuc = "G"}; for a cytosine base editor, \code{from_nuc = "C"} and \code{to_nuc = "T"}.

Note that \code{n_reads_w_modification} is not simply the sum over \code{n_reads_w_intended_substitution}, \code{n_reads_w_bystander_substitution}, and \code{n_reads_w_indel}, as some reads may harbor multiple modifications of different types (e.g., an intended substitution and an indel).
}
\examples{
crispresso_dir <- "/Users/timbarry/research_offsite/external/scratch/novaseq_12_elane_mrna_dose_response/chr19_853252_853275/crispresso_outputs/CRISPResso_on_Jing_i507_i733_on_target_A8toG/"
count_df <- compute_base_editing_count_table(crispresso_dir = crispresso_dir,
                                             bystander_substitution_positions = 10L,
                                             intended_substitution_positions = 8L)
}
