% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_amplicon_seq.R
\name{compute_basic_statistics_on_multivariate_amplicon_seq_data}
\alias{compute_basic_statistics_on_multivariate_amplicon_seq_data}
\title{Compute basic statistics on multivariate amplicon-seq data}
\usage{
compute_basic_statistics_on_multivariate_amplicon_seq_data(
  count_matrix,
  covariate_df
)
}
\arguments{
\item{count_matrix}{the allele count matrix}

\item{covariate_df}{the allele covariate data frame}
}
\value{
a data frame with columns \code{editing_rate} and \code{attributable_fraction}
}
\description{
This function computes the editing rate and attributable fraction of each allele in a multivariate allele table.
}
\examples{
data_list <- readRDS("/Users/timbarry/research_offsite/external/bauer-lab/rhampseq_bcl11a/multivariate_count_tables/combined/CRISPResso_on_1617_OT_0068.rds")
count_matrix <- data_list$count_matrix
covariate_df <- data_list$covariate_df
}
