% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_model.R
\name{compute_dm_log_lik}
\alias{compute_dm_log_lik}
\title{Compute DM log likelihood}
\usage{
compute_dm_log_lik(r, pi, theta)
}
\arguments{
\item{r}{the vector of counts}

\item{pi}{the mean of the DM model}

\item{theta}{the size parameter of the DM model}
}
\value{
the log-likelihood (up to the multinomial coefficient normalizing constant)
}
\description{
Computes the DM log-likelihood for a single vector of counts
}
