% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicate_aware_guide_seq.R
\name{construct_replicate_count_table}
\alias{construct_replicate_count_table}
\title{Construct replicate count table}
\usage{
construct_replicate_count_table(
  clustered_count_df,
  channel_axes = c("replicate_id", "primer_type")
)
}
\arguments{
\item{clustered_count_df}{output of \code{cluster_loci()}, with columns \code{window}, \code{umi_count}, and the columns specified by \code{channel_axes}.}

\item{channel_axes}{columns whose unique tuples define evidence channels.}
}
\value{
an integer matrix with evidence channels in the rows and windows in the columns. An entry of the matrix indicates the number of UMIs observed within a given window and channel.
}
\description{
Takes the output of \code{cluster_loci()} as input; outputs a count matrix for statistical modeling
}
\examples{
elane_dir <- paste0(.get_config_path("LOCAL_BAUER_LAB_DATA_DIR"), "guideseq_elane/")
count_df <- readRDS(paste0(elane_dir, "count_tables_no_multimap/combined_count_df.rds")) |>
 dplyr::filter(cell_type == "CD34" & cas9_variant == "wt_cas9" & treated & replicate_id \%in\% 1:2, chr != "chrM") |>
 dplyr::select(chr, coord, strand, umi_count, primer_type, replicate_id)
clustered_count_df <- cluster_loci(count_df)
Y_mat <- construct_replicate_count_table(clustered_count_df)
}
