% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplicon_seq_power_analysis.R
\name{design_n_and_r_amplicon_seq}
\alias{design_n_and_r_amplicon_seq}
\title{Design n and r in an amplicon-seq experiment}
\usage{
design_n_and_r_amplicon_seq(
  pi_cntrl,
  theta,
  rho,
  ci_width = 0.01,
  alpha = 0.05
)
}
\arguments{
\item{pi_cntrl}{mutation rate in control condition}

\item{theta}{editing rate}

\item{rho}{overdispersion parameter}

\item{ci_width}{width of the confidence interval}

\item{alpha}{the two-sided confidence interval is level (1-alpha)}
}
\description{
Given key parameters of an amplicon-seq study, design n (number of reads) and r (number of independent samples)
}
\examples{
pi_cntrl <- 0.03
theta <- 0.05
rho <- 0.00025
out <- design_n_and_r_amplicon_seq(pi_cntrl, theta, rho)
}
