% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_outlier_detection.R
\name{find_guideseq_edit_sites_dm_clustered}
\alias{find_guideseq_edit_sites_dm_clustered}
\title{Find GUIDE-seq edit sites DM clustered}
\usage{
find_guideseq_edit_sites_dm_clustered(
  count_df,
  max_chain_link = 25,
  multiplicity_adjustment = "BH",
  multiplicity_alpha = 0.1,
  chrs_to_keep = seq(1L, 22L),
  fit = NULL,
  plot_col = c("dodgerblue3", "firebrick")[1]
)
}
\arguments{
\item{plot_col}{}
}
\description{
Find GUIDE-seq edit sites DM clustered
}
\examples{
data_dir <- .get_config_path("LOCAL_CRISPR_DE_DATA_DIR")

umi_tab_fp <- paste0(data_dir, "/guideseq/count_tables/293T-SpRY-Cas9-dsODN-only-GSPneg-S81_S89_L001_count_table.rds")
count_df <- readRDS(umi_tab_fp)
res_cntrl <- find_guideseq_edit_sites_dm_clustered(count_df, plot_col = "firebrick")

umi_tab_fp <- paste0(data_dir, "guideseq/count_tables/293T-SpRY-Cas9-1620-GSPneg-S79_S87_L001_count_table.rds")
count_df <- readRDS(umi_tab_fp)
res_trt <- find_guideseq_edit_sites_dm_clustered(count_df)
}
