% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_outlier_detection.R
\name{find_guideseq_edit_sites_dm_sliding_window}
\alias{find_guideseq_edit_sites_dm_sliding_window}
\title{Find GUIDE-seq edit sites DM}
\usage{
find_guideseq_edit_sites_dm_sliding_window(
  count_df,
  window_size = 25,
  multiplicity_adjustment = "BH",
  multiplicity_alpha = 0.1,
  chrs_to_keep = seq(1L, 22L),
  fit = NULL
)
}
\arguments{
\item{plot_col}{}
}
\description{
Find GUIDE-seq edit sites DM
}
\examples{
umi_tab_fp <- "/Users/timbarry/research_offsite/external/crispr-quant/guideseq/count_tables/293T-SpRY-Cas9-dsODN-only-GSPneg-S81_S89_L001_count_table.rds"
count_df <- readRDS(umi_tab_fp)
res <- find_guideseq_edit_sites_dm_sliding_window(count_df)
}
