% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_model.R
\name{fit_dm_model}
\alias{fit_dm_model}
\title{Fit DM model}
\usage{
fit_dm_model(X)
}
\arguments{
\item{X}{A matrix of dimension n by k, where n is the number of samples and k the number of categories}
}
\value{
a fitted DM model
}
\description{
Fits a DM model to a matrix of counts X.
}
\examples{
n_umis_in_window <- 500
n <- 1000
K <- 5
pi <- c(0.1, 0.2, 0.4, 0.2, 0.1)
X <- dirmult::simPop(J = n, K = K, n = n_umis_in_window, pi = pi, theta = 0.05)$data
fit <- fit_dm_model(X)
pi_hat <- fit$pi
theta_hat <- fit$theta
p_vals <- compute_lrt_p_vals(X = X, pi = pi_hat, theta = theta_hat)
p_vals <- compute_lrt_p_vals_v2(X = X, pi = pi_hat, theta = theta_hat)
}
