% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_amplicon_seq.R
\name{fit_multivariate_bayesian_regression_model}
\alias{fit_multivariate_bayesian_regression_model}
\title{Fit multivariate Bayesian regression model}
\usage{
fit_multivariate_bayesian_regression_model(
  count_matrix,
  allele_df,
  covariate_df,
  rho_hat,
  formula_object,
  mu_pi_block,
  kappa_pi_block,
  mu_theta,
  kappa_theta,
  mu_phi_block,
  kappa_phi_block,
  mu_gamma,
  sigma_gamma,
  mu_delta,
  sigma_delta
)
}
\description{
Fit multivariate Bayesian regression model
}
\examples{
# prepare data input
data_list <- readRDS("/Users/timbarry/research_offsite/external/bauer-lab/rhampseq_bcl11a/multivariate_count_tables/combined/CRISPResso_on_1450_OT_0000.rds")
count_matrix <- data_list$count_matrix
allele_df <- data_list$allele_df
covariate_df <- data_list$covariate_df
collapsed_matrix_list <- collapse_count_matrix_by_allele(count_matrix, allele_df)
count_matrix <- collapsed_matrix_list$collapsed_matrix
allele_df <- collapsed_matrix_list$collapsed_allele_df
# for now -- restrict attention to deletions
to_keep <- allele_df$mutation_type \%in\% c(NA, "deletion")
count_matrix <- count_matrix[,to_keep]
allele_df <- allele_df[to_keep,]

# get rho from univariate analysis
dat_list <- readRDS("/Users/timbarry/research_offsite/external/bauer-lab/rhampseq_bcl11a/marginal_count_tables/data_list_by_grna.rds")[["1450"]]
univariate_res <- run_freqentist_amplicon_seq_analysis(data_list = dat_list)
rho_hat <- univariate_res$result_df |> dplyr::filter(amplicon_id == "1450_OT_0000") |> dplyr::pull(rho_hat)

# formula object
formula_object <- formula(~ mutation_length + 0)

# hyperparameters
# pi (control mean)
mu_pi_block <- c(0.999, 0.001)
kappa_pi_block <- 100

# theta (editing rate)
mu_theta <- 0.5
kappa_theta <- 20

# block level editing
mu_phi_block <- 1
kappa_phi_block <- 1

# control side regression
mu_gamma <- matrix(-1)
sigma_gamma <- matrix(0.5)

# treatment side regression
mu_delta <- matrix(-1)
sigma_delta <- matrix(0.5)

# fit model
fit <- fit_multivariate_bayesian_regression_model(count_matrix, allele_df, covariate_df, rho_hat, formula_object,
mu_pi_block, kappa_pi_block,
mu_theta, kappa_theta,
mu_phi_block, kappa_phi_block,
mu_gamma, sigma_gamma,
mu_delta, sigma_delta)
}
