% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicate_aware_guide_seq.R
\name{fit_tbp_model}
\alias{fit_tbp_model}
\title{Fit truncated Bernoulli product (TBP) model via MLE}
\usage{
fit_tbp_model(X)
}
\arguments{
\item{X}{binary occupancy matrix of dimension r (number of replicates) by m (number of windows)}
}
\value{
the fitted MLE pi-hat (or NULL if the regularity conditions fail to hold)
}
\description{
Fits a TBP model to an r x m matrix of occupancies.
}
\details{
Returns NULL if the regularity conditions fail to hold.
}
\examples{
m <- 5000L
pi <- c(0.2, 0.12, 0.04)
X <- r_tbp(m, pi)
pi_hat <- fit_tbp_model(X)
}
