% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplicon_seq.R
\name{generate_synthetic_amplicon_seq_data}
\alias{generate_synthetic_amplicon_seq_data}
\title{Generate synthetic amplicon-seq data}
\usage{
generate_synthetic_amplicon_seq_data(
  p,
  r,
  pi_cntrl,
  editing_rate,
  n_amplicons_nonzero_editing,
  beta_binom_rho,
  amplicon_ids,
  sample_size_mu = 100000L,
  sample_size_theta = 15L
)
}
\arguments{
\item{p}{number of amplicons}

\item{r}{number of treatment/control samples}

\item{pi_cntrl}{(\link{0,1} scalar) background editing rate to use across all control samples}

\item{editing_rate}{(\link{0,1} scalar) editing rate in samples where there is nonzero editing}

\item{n_amplicons_nonzero_editing}{(integer) number of amplicons}

\item{beta_binom_rho}{(positive scalar or vector) dispersion parameter for the beta-binomial distribution}

\item{amplicon_ids}{vector of amplicon identifiers}

\item{sample_size_mu}{(positive scalar) mean sample size}

\item{sample_size_theta}{(positive scalar) size (i.e., overdispersion) parameter for sample size}
}
\value{
A list containing the simulated treated/control count matrices, amplicon identifiers,
the generating mutation rates, and the beta-binomial dispersion values used in the simulation.
}
\description{
Generate synthetic amplicon-seq data
}
\examples{
p <- 50L
sim_data <- generate_synthetic_amplicon_seq_data(
  p = 50L,
  r = 3L,
  pi_cntrl = 0.02,
  editing_rate = 0.15,
  n_amplicons_nonzero_editing = 2L,
  beta_binom_rho = 0.005,
  amplicon_ids = paste0("amplicon_", seq_len(p)),
  sample_size_mu = 50L
)
}
