% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_data.R
\name{generate_synthetic_multivariate_amplicon_seq_data}
\alias{generate_synthetic_multivariate_amplicon_seq_data}
\title{Generate synthetic multivariate amplicon-seq data}
\usage{
generate_synthetic_multivariate_amplicon_seq_data(
  allele_feature_table,
  use_observed_n_reads = FALSE,
  sample_size_mu = 100000L,
  sample_size_theta = 15L,
  frac_mutated = 0.8,
  n_rep = 3L,
  rho = 2e-05,
  beta_mut_length = -0.1
)
}
\arguments{
\item{allele_feature_table}{allele}

\item{frac_mutated}{fraction of reads containing a mutation}

\item{rho}{overdispersion parameter}

\item{beta_mut_length}{model parameter controlling the relationship between mutation length and allele frequency}

\item{n_reads}{number of reads (fixed across replicates)}
}
\value{
a count table containing the data
}
\description{
Generate synthetic, multivariate amplicon-seq data using an example allele table as input.
}
\examples{
allele_table_fp <- "/Users/timbarry/research_offsite/external/bauer-lab/rhampseq_bcl11a/crispresso_output_1617/GUIDE0724-africa-edited/CRISPResso_on_1617_OT_0000/Alleles_frequency_table_around_sgRNA_CTAACAGTTGCTTTTATCAC.txt"
allele_feature_table <- process_crispresso_allele_table_cas9(allele_table_fp)
count_tab <- generate_synthetic_multivariate_amplicon_seq_data(allele_feature_table, n_reads = 100000L)
}
