% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_management.R
\name{get_trt_cntrl_tab}
\alias{get_trt_cntrl_tab}
\title{data_dir <- paste0(.get_config_path("LOCAL_CRISPR_DE_DATA_DIR"), "guideseq/count_tables/")}
\usage{
get_trt_cntrl_tab(data_dir)
}
\description{
data_dir <- paste0(.get_config_path("LOCAL_CRISPR_DE_DATA_DIR"), "guideseq/count_tables/")
}
