% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicate_aware_guide_seq.R
\name{load_crispritz_output}
\alias{load_crispritz_output}
\title{Load CRISPRitz output}
\usage{
load_crispritz_output(targets_file_path)
}
\arguments{
\item{targets_file_path}{file path to the targets.txt file outputted by CRISPRitz}
}
\value{
a data frame containing the CRISPRitz target output
}
\description{
Load CRISPRitz output
}
\examples{
homology_df <- load_crispritz_output("/Users/timbarry/research_offsite/external/bauer-lab/guideseq_elane/crispritz/crispritz_CCCCGGCAGAAACGTCCGCG.hg38.targets.txt")
}
