% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicate_aware_guide_seq.R
\name{load_encode_blacklist_bed}
\alias{load_encode_blacklist_bed}
\title{Load ENCODE blacklist bed}
\usage{
load_encode_blacklist_bed(encode_blacklist_file_path)
}
\arguments{
\item{encode_blacklist_file_path}{filepath to ENCODE blacklist}
}
\value{
a data frame storing the positions of the blacklisted regions
}
\description{
Load ENCODE blacklist bed
}
\examples{
encode_blacklist_file_path <- "/Users/timbarry/research_offsite/ref_genome_dir/hg38-blacklist.v2.bed"
}
