% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicate_aware_guide_seq.R
\name{load_n_run_bed}
\alias{load_n_run_bed}
\title{Load N run bed file}
\usage{
load_n_run_bed(n_run_bed_file_path)
}
\arguments{
\item{n_run_bed_file_path}{path to N-run bed file}
}
\value{
a data frame storing the positions of the N-runs
}
\description{
Load bed file storing the positions of N-runs in the reference genome.
}
\examples{
n_run_bed_file_path <- "/Users/timbarry/research_offsite/ref_genome_dir/hg38_N_runs_min10.bed"
n_run_df <- load_n_run_bed(n_run_bed_file_path)
}
