% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_pt_3.R
\name{make_guideseq_qq_plot}
\alias{make_guideseq_qq_plot}
\title{Make GUIDE-seq QQ plot}
\usage{
make_guideseq_qq_plot(
  res_df,
  color_ground_truth = FALSE,
  rev_log_trans = TRUE,
  min_p = 1e-16,
  annotate_discoveries = FALSE,
  point_col = "black",
  rej_threshold_col = "blue",
  annotation_size = 1.5,
  annotation_color = "red"
)
}
\arguments{
\item{res_df}{a data frame with column \code{p_value} giving the p-value of a given locus}
}
\value{
a qq plot of the results
}
\description{
Make GUIDE-seq QQ plot
}
