% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_pt_3.R
\name{make_local_scatterplot}
\alias{make_local_scatterplot}
\title{Make local scatterplot}
\usage{
make_local_scatterplot(annotated_df_sub, title = NULL)
}
\arguments{
\item{annotated_df_sub}{the annotated data frame for one window. gRNA and DNA sequence information must be present.}

\item{title}{plot title}
}
\value{
a ggplot of the local UMI count distribution
}
\description{
Make local scatterplot
}
\examples{
caliper_res <- readRDS("/Users/timbarry/research_offsite/projects/crisprde-project/guideseq/caliper_res_list_with_p_val_weighting.rds")
res_df <- caliper_res$bcl11a_293t_1620_sprycas9$tuning_res$selected_trt_run$res_df
annotated_clustered_count_df_trt <- caliper_res$bcl11a_293t_1620_sprycas9$annotated_clustered_count_df_trt

# minus strand PAM
window_id <- res_df |> dplyr::slice(1L) |> dplyr::pull(window)
annotated_df_sub <- annotated_clustered_count_df_trt |> dplyr::filter(window == window_id)
p <- make_local_scatterplot(annotated_df_sub)

# "-" in rna alignment, minus starnd PAM
window_id <- "chr10:130663775-130663848"
annotated_df_sub <- annotated_clustered_count_df_trt |> dplyr::filter(window == window_id)
p <- make_local_scatterplot(annotated_df_sub)

# "-" in dna alignment, minus strand PAM
window_id <- "chr3:3330681-3330751"
annotated_df_sub <- annotated_clustered_count_df_trt |> dplyr::filter(window == window_id)
p <- make_local_scatterplot(annotated_df_sub)

# "-" in dna alignment, plus strand PAM
window_id <- "chr2:8943413-8943485"
annotated_df_sub <- annotated_clustered_count_df_trt |> dplyr::filter(window == window_id)
p <- make_local_scatterplot(annotated_df_sub)

# plus strand pam
window_id <- "chr5:56529490-56529575"
annotated_df_sub <- annotated_clustered_count_df_trt |> dplyr::filter(window == window_id)
p <- make_local_scatterplot(annotated_df_sub)
}
