% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_bayesian_amplicon_seq.R
\name{make_prior_density_plot}
\alias{make_prior_density_plot}
\title{Plot a Beta prior density for Bayesian amplicon-seq analysis}
\usage{
make_prior_density_plot(
  alpha,
  beta,
  parameter = c("theta", "pi"),
  n_grid = 1000L,
  line_color = "dodgerblue3",
  x_limits = c(0, 1)
)
}
\arguments{
\item{alpha}{First shape parameter of the Beta prior.}

\item{beta}{Second shape parameter of the Beta prior.}

\item{parameter}{Which parameter prior to plot, either \code{"theta"} or \code{"pi"}.}

\item{n_grid}{Number of grid points used to evaluate the density on \verb{(0, 1)}.}

\item{line_color}{Color of the density curve.}

\item{xmin}{Lower endpoint used to evaluate the density grid.}

\item{xmax}{Upper endpoint used to evaluate the density grid.}
}
\value{
A \code{ggplot2} object showing the requested Beta prior density.
}
\description{
Draw the probability density function of the Beta prior used for either the
background mutation rate \code{pi} or the editing rate \code{theta}.
}
\examples{
make_prior_density_plot(parameter = "theta", alpha = 1, beta = 1)
make_prior_density_plot(parameter = "pi", alpha = 2, beta = 50, line_color = "firebrick")
}
