% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{make_scatterplot}
\alias{make_scatterplot}
\title{Make scatterplot}
\usage{
make_scatterplot(
  count_df,
  x_range = NULL,
  facet_on_chr = FALSE,
  log_trans = FALSE,
  point_size = 1,
  col = c("dodgerblue3", "firebrick")[1],
  title = NULL,
  target_seq = NULL,
  grna_spacer = NULL,
  pam_strand = NULL,
  protospacer_range = NULL,
  pam_range = NULL,
  cut_base_range = NULL
)
}
\arguments{
\item{cut_base_range}{}
}
\description{
Make scatterplot
}
