% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicate_aware_guide_seq.R
\name{overlap_homology_and_result_dfs}
\alias{overlap_homology_and_result_dfs}
\title{Overlap homology and result df}
\usage{
overlap_homology_and_result_dfs(result_df, homology_df)
}
\arguments{
\item{result_df}{output of run_multireplicate_guideseq_method}

\item{homology_df}{a homology data frame (for instance, output of load_crispritz_output)}
}
\value{
an augmented version of \code{result_df} with homology information appended
}
\description{
Overlap homology and result df
}
