% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_pt_2.R
\name{plot_read_processing_results}
\alias{plot_read_processing_results}
\title{Plot read processing results}
\usage{
plot_read_processing_results(read_proc_df)
}
\arguments{
\item{read_proc_df}{the output of \code{run_read_qc_on_sample()}}
}
\description{
Plot read processing results
}
\examples{
sample_dir <- "/Users/timbarry/research_code/genethoff-nf/demo/results/Jing_AAVS1_n1-10_Donor-Seq_AAVS1_GSP_plus_1"
read_proc_df <- run_read_qc_on_sample(sample_dir)
plot_read_processing_results(read_proc_df)
}
