% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicate_aware_guide_seq.R
\name{pmf_tbp}
\alias{pmf_tbp}
\title{Returns the pmf of a truncated Bernoulli product (TBP) distribution}
\usage{
pmf_tbp(pi)
}
\arguments{
\item{pi}{parameter of the TBP distribution}
}
\value{
a data frame with columns (x_1, x_2, \dots, x_r, pmf), where pmf gives the probability of a given (x_1, x_2, \dots, x_r) vector
}
\description{
Returns the pmf of a truncated Bernoulli product (TBP) distribution
}
\examples{
Omega <- pmf_tbp(c(0.1, 0.05, 0.02))
Omega <- pmf_tbp(c(0.4, 0.6, 0.2))
}
