% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicate_aware_guide_seq.R
\name{r_snb}
\alias{r_snb}
\title{shifted negative binomial (SNB) distribution}
\usage{
r_snb(m_plus, mu, theta)
}
\arguments{
\item{m_plus}{the number of samples to generate (typically equal to the number of occupied windows)}

\item{mu}{mean parameter}

\item{theta}{size parameter}
}
\value{
a vector of snb variates
}
\description{
Sample m_plus observations from an SNB distribution with parameters (mu, theta)
}
\examples{
mu <- 10
theta <- 0.5
m_plus <- 1000L
y_plus <- r_snb(m_plus, mu, theta)
}
