% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicate_aware_guide_seq.R
\name{r_tbp}
\alias{r_tbp}
\title{Truncated Bernoulli product (TBP) distribution}
\usage{
r_tbp(m, pi)
}
\arguments{
\item{m}{number of samples to draw}

\item{pi}{the TBP parameter vector of length r}
}
\value{
a matrix of dimension r x m of TBP draws
}
\description{
Simulate draws from a TBP distribution
}
\examples{
m <- 5000L
pi <- c(0.1, 0.05, 0.1)
X <- r_tbp(m, pi)
}
