% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicate_aware_guide_seq.R
\name{run_multivariate_guideseq_method}
\alias{run_multivariate_guideseq_method}
\title{Run multivariate guide-seq method}
\usage{
run_multivariate_guideseq_method(
  Y_mat,
  incorporate_occupancy_info = TRUE,
  multiplicity_alpha = 0.1
)
}
\arguments{
\item{Y_mat}{the r x m integer matrix of UMI counts, where r is the number of replicates and m is the number of windows}

\item{incorporate_occupancy_info}{a boolean (T/F) indicating whether to incorporate occupancy information into the p-value calculation}
}
\value{
a data frame containing a p-value for each window
}
\description{
Runs the multivariate guide-seq method on a replicate-by-window matrix of counts
}
\examples{
# NULL DATA
pi <- c(0.05, 0.1, 0.02)
mu_vect <- c(10, 6, 15)
theta_vect <- c(2, 5, 0.3)
m <- 10000
null_dat <- simulate_multirep_guideseq_data(pi, mu_vect, theta_vect, m)

# ALTERNATIVE DATA
pi <- c(0.5, 0.6, 0.4)
mu_vect <- c(80, 200, 50)
theta_vect <- c(20, 21, 15)
m <- 15
alt_dat <- simulate_multirep_guideseq_data(pi, mu_vect, theta_vect, m)

# COMBINED DATA
Y_mat <- cbind(alt_dat, null_dat)
incorporate_occupancy_info <- TRUE
}
