% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicate_aware_guide_seq.R
\name{simulate_multirep_guideseq_data}
\alias{simulate_multirep_guideseq_data}
\title{Simulate multi-replicate guide-seq data}
\usage{
simulate_multirep_guideseq_data(pi, mu_vect, theta_vect, m)
}
\arguments{
\item{pi}{parameter vector of TBP model}

\item{mu_vect}{vector of mu parameters for SNB models}

\item{theta_vect}{vector of theta parameters for SNB models}
}
\description{
Simulate multi-replicate guide-seq data
}
\examples{
# NULL DATA
pi <- c(0.05, 0.1, 0.02)
mu_vect <- c(10, 6, 15)
theta_vect <- c(2, 5, 0.3)
m <- 10000
null_dat <- simulate_multirep_guideseq_data(pi, mu_vect, theta_vect, m)

# ALTERNATIVE DATA
pi <- c(0.5, 0.6, 0.4)
mu_vect <- c(80, 200, 50)
theta_vect <- c(20, 21, 15)
m <- 15
alt_dat <- simulate_multirep_guideseq_data(pi, mu_vect, theta_vect, m)

# COMBINED DATA
Y_mat <- cbind(alt_dat, null_dat)
colnames(Y_mat) <- paste0("window_", seq_len(ncol(Y_mat)))
}
