% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicate_aware_guide_seq.R
\name{annotate_clustered_count_df}
\alias{annotate_clustered_count_df}
\title{Annotate clustered count df with homology}
\usage{
annotate_clustered_count_df(
  clustered_count_df,
  homology_df = NULL,
  n_run_df = NULL,
  encode_blacklist_df = NULL
)
}
\arguments{
\item{clustered_count_df}{output of \code{cluster_loci()}}

\item{homology_df}{optional output of \code{load_crispritz_output()}; if supplied, windows are annotated for overlap with CRISPRitz hits}

\item{n_run_df}{optional output of \code{load_n_run_bed()}; if supplied, windows are annotated for overlap with N-runs}

\item{encode_blacklist_df}{optional output of \code{load_encode_blacklist_bed()}; if supplied, windows are annotated for overlap with ENCODE blacklist regions}
}
\description{
Add columns:
\itemize{
\item \code{homology_has_hit} (indicating whether there is a homology hit inside the window)
\item \code{overlaps_n_run} (indicating whether the window overlaps an N-run in the reference genome)
\item \code{overlaps_encode_blacklist} (indicating whether the window overlaps an ENCODE blacklist region)
\item \code{homology_n_mismatches} (indicating the number of mismatches between aligned spacer and protospacer sequence)
\item \code{homology_n_bulges} (indicating number of bulges between aligned spacer and protospacer sequence)
\item \code{homology_bulge_type} (indicating the CRISPRitz bulge type)
\item \code{homology_cfd} (indicating the CFD-like homology score)
\item \code{homology_posit} (indicating the position of the start of the protospacer)
\item \code{homology_strand} (indicating whether the protospacer is on the plus or minus strand)
\item \code{homology_dna} (aligned protospacer sequence)
\item \code{homology_gRNA} (aligned spacer sequence)
}
}
\examples{
homology_df <- load_crispritz_output("/Users/timbarry/research_offsite/external/bauer-lab/guideseq_bcl11a/1620_crispritz_spRY_bcl11a_windows.hg38.targets.txt")
n_run_df <- load_n_run_bed("/Users/timbarry/research_offsite/ref_genome_dir/hg38_N_runs_min10.bed")
encode_blacklist_df <- load_encode_blacklist_bed("/Users/timbarry/research_offsite/ref_genome_dir/hg38-blacklist.v2.bed")
bcl11a_dir <- paste0(.get_config_path("LOCAL_BAUER_LAB_DATA_DIR"), "guideseq_bcl11a/")
clustered_count_df <- readRDS(paste0(bcl11a_dir, "count_tables_no_multimap/combined_count_df.rds")) |>
dplyr::filter(grna == "1620" & treated & chr != "chrM") |>
dplyr::select(chr, coord, strand, umi_count, primer_type, replicate_id) |>
cluster_loci()
annotated_clustered_count_df <- annotate_clustered_count_df(clustered_count_df = clustered_count_df,
  homology_df = homology_df, n_run_df = n_run_df, encode_blacklist_df = encode_blacklist_df)
}
