% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boost_p_values.R
\name{boost_p_values_genovese}
\alias{boost_p_values_genovese}
\title{Boost p-values (Genovese)}
\usage{
boost_p_values_genovese(
  augmented_result_df,
  multiplicity_alpha = 0.5,
  gamma_align = NULL,
  gamma_distance = NULL
)
}
\arguments{
\item{augmented_result_df}{result data frame with homology annotations}

\item{multiplicity_alpha}{nominal FDR}

\item{gamma_align}{tuning parameter controlling the homology-score penalty}

\item{gamma_distance}{tuning parameter controlling the modal-base-distance penalty}
}
\description{
Boost p-values (Genovese)
}
\examples{
elane_dir <- paste0(.get_config_path("LOCAL_BAUER_LAB_DATA_DIR"), "guideseq_elane/")
clustered_count_df <- readRDS(paste0(elane_dir, "count_tables_no_multimap/combined_count_df.rds")) |>
 dplyr::filter(cell_type == "CD34" & cas9_variant == "wt_cas9" & treated & replicate_id \%in\% 1:2 & chr != "chrM") |>
 dplyr::select(chr, coord, strand, umi_count, replicate_id) |>
 cluster_loci()
homology_df <- load_crispritz_output("/Users/timbarry/research_offsite/external/bauer-lab/guideseq_elane/crispritz_CCCCGGCAGAAACGTCCGCG.hg38.targets.txt")
annotated_clustered_count_df <- annotate_clustered_count_df(clustered_count_df, homology_df) # |> dplyr::filter(homology_has_hit)
Y_mat <- construct_replicate_count_table(annotated_clustered_count_df)
augmented_result_df <- run_multireplicate_guideseq_method(Y_mat = Y_mat, lambda = 10, c_tukey_sigma = 50, multiplicity_alpha = 0.2, robust_fit = TRUE, incorporate_occupancy_info = TRUE, annotated_clustered_count_df = annotated_clustered_count_df)$res_df

# compute alignment scores/weights
alignment_scores <- compute_alignment_scores(
  cfds = augmented_result_df$homology_cfd,
  distances = augmented_result_df$homology_modal_base_cut_distance,
  has_hits = augmented_result_df$homology_has_hit
  )

# substitution/bulge/distance weighting
weighted_result_df <- boost_p_values_genovese(augmented_result_df)
qq_plot <- weighted_result_df |> make_guideseq_qq_plot()

# cfd weighting
weighted_result_df <- boost_p_values_genovese_cfd(augmented_result_df)
qq_plot <- weighted_result_df |> make_guideseq_qq_plot()
}
