% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boost_p_values.R
\name{boost_p_values_genovese}
\alias{boost_p_values_genovese}
\title{Boost p-values (Genovese)}
\usage{
boost_p_values_genovese(
  augmented_result_df,
  multiplicity_alpha = 0.2,
  gamma = 0.5
)
}
\arguments{
\item{augmented_result_df}{output of \code{overlap_homology_and_result_dfs}}

\item{multiplicity_alpha}{nominal FDR}

\item{gamma}{gamma}
}
\description{
Boost p-values (Genovese)
}
\examples{
elane_dir <- paste0(.get_config_path("LOCAL_BAUER_LAB_DATA_DIR"), "guideseq_elane/")
count_df <- readRDS(paste0(elane_dir, "count_tables_no_multimap/combined_count_df.rds")) |>
 dplyr::filter(cell_type == "CD34" & cas9_variant == "wt_cas9" & treated & replicate_id \%in\% 1:2) |>
 dplyr::filter(chr != "chrM") |>
 dplyr::select(chr, coord, strand, umi_count, primer_type, replicate_id)
Y_mat <- construct_replicate_count_table(count_df)
result_df <- run_multireplicate_guideseq_method(Y_mat = Y_mat, lambda = 10, c_tukey_sigma = 50, multiplicity_alpha = 0.2, robust_fit = TRUE, incorporate_occupancy_info = TRUE)$res_df
homology_df <- load_crispritz_output("/Users/timbarry/research_offsite/external/bauer-lab/guideseq_elane/crispritz/crispritz_CCCCGGCAGAAACGTCCGCG.hg38.targets.txt")
augmented_result_df <- overlap_homology_and_result_dfs(result_df, homology_df)
weighted_result_df <- boost_p_values_genovese(augmented_result_df)
qq_plot <- weighted_result_df |> dplyr::mutate(p_value = p_value_weighted) |> make_guideseq_qq_plot()
}
