% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicate_aware_guide_seq.R
\name{cluster_loci}
\alias{cluster_loci}
\title{Cluster loci}
\usage{
cluster_loci(count_df, thresh = 100L, padding = 25L)
}
\arguments{
\item{count_df}{a data frame containing columns chr, coord}

\item{thresh}{clustering threshold; occupied bases within this distance are clustered together}

\item{padding}{amount of padding to add to either side of each cluster}
}
\value{
\code{count_df} with the additional columns appended:
\itemize{
\item \code{group_string} (a string indicating the cluster to which a given base belongs)
\item \code{group_chr} (chromosome of the group)
\item \code{min_group_coord} (minimum coordinate of the group)
\item \code{max_group_coord} (maximum coordinate of the group)
}
}
\description{
Clusters loci via single-linkage clustering
}
\examples{
elane_dir <- paste0(.get_config_path("LOCAL_BAUER_LAB_DATA_DIR"), "guideseq_elane/")
count_df <- readRDS(paste0(elane_dir, "count_tables_no_multimap/combined_count_df.rds")) |>
 dplyr::filter(cell_type == "CD34" & cas9_variant == "wt_cas9" & treated & replicate_id \%in\% 1:2) |>
 dplyr::filter(chr != "chrM") |>
 dplyr::select(chr, coord, strand, umi_count, primer_type, replicate_id)
clustered_count_df <- cluster_loci(count_df)
}
