% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_amplicon_seq.R
\name{compute_bayesian_credible_interval}
\alias{compute_bayesian_credible_interval}
\title{Title}
\usage{
compute_bayesian_credible_interval(
  n_trt,
  n_cntrl,
  k_trt,
  k_cntrl,
  rho,
  alpha_pi_cntrl,
  beta_pi_cntrl,
  alpha_theta,
  beta_theta,
  nominal_ci_probability = 0.95
)
}
\arguments{
\item{nominal_ci_probability}{}
}
\description{
Title
}
\examples{
p <- 5L
amplicon_ids <- factor(x = paste0("amplicon_", seq_len(p)), levels = paste0("amplicon_", seq_len(p)))
beta_binom_rho <- rep(5e-4, times = p)
data_list <- generate_synthetic_amplicon_seq_data(p = p, r = 3L, pi_cntrl = 0.05, editing_rate = 0.15,
                                                  n_amplicons_nonzero_editing = 2L, beta_binom_rho,
                                                  amplicon_ids = amplicon_ids)
n_trt <- data_list$n_mat_trt[,1]
n_cntrl <- data_list$n_mat_cntrl[,1]
k_trt <- data_list$k_mat_trt[,1]
k_cntrl <- data_list$k_mat_cntrl[,1]
rho <- beta_binom_rho[1]
alpha_pi_cntrl <- 1
beta_pi_cntrl <- 100
alpha_theta <- 1
beta_theta <- 100
}
