% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicate_aware_guide_seq.R
\name{construct_replicate_count_table}
\alias{construct_replicate_count_table}
\title{Construct replicate count table}
\usage{
construct_replicate_count_table(count_df, thresh = 100L, padding = 25L)
}
\arguments{
\item{count_df}{a data frame with columns \code{chr}, \code{coord}, \code{strand}, \code{replicate_id}, and \code{umi_count}.}

\item{thresh}{clustering threshold; bases within this distance are clustered together into a window}

\item{padding}{amount of padding to add to either end of each window}
}
\value{
an integer matrix with replicates in the rows and windows in the columns. An entry of the matrix indicates the number of UMIs observed within a given window and replicate.
}
\description{
Takes the output from the GUIDE-seq pipeline as input; outputs a count matrix for statistical modeling
}
\examples{
elane_dir <- paste0(.get_config_path("LOCAL_BAUER_LAB_DATA_DIR"), "guideseq_elane/")
count_df_trt <- readRDS(paste0(elane_dir, "count_tables_no_multimap/combined_count_df.rds")) |>
 dplyr::filter(cell_type == "CD34" & cas9_variant == "wt_cas9" & treated & replicate_id \%in\% 1:2) |>
 dplyr::filter(chr != "chrM")
Y_mat <- construct_replicate_count_table(count_df_trt)
res <- run_multireplicate_guideseq_method(Y_mat = Y_mat, lambda = 10)
}
