% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nb_outlier_detection.R
\name{find_guideseq_edit_sites}
\alias{find_guideseq_edit_sites}
\title{Find GUIDE-seq edit sites}
\usage{
find_guideseq_edit_sites(
  count_df,
  grna_spacer = NULL,
  pam = NULL,
  bin_genome = FALSE,
  window_size = 1000,
  max_chain_link = 50,
  multiplicity_adjustment = "BH",
  fit = NULL,
  use_offset = FALSE,
  model = "nb",
  robust_mle = TRUE,
  multiplicity_alpha = 0.1,
  c_tukey_beta = 10,
  c_tukey_sigma = 10,
  chrs_to_keep = paste0("chr", seq(1L, 22L)),
  plot_col = c("dodgerblue3", "firebrick")[1]
)
}
\arguments{
\item{chrs_to_keep}{}
}
\description{
TO DO:
\enumerate{
\item optimize c_tukey_beta and c_tukey_sigma for our setting
\item consider binning-free approaches
\item Multiple samples and covariates
}
}
\examples{
data_dir <- paste0(.get_config_path("LOCAL_BAUER_LAB_DATA_DIR"), "guideseq_elane/")
combined_count_df <- readRDS(paste0(data_dir, "count_tables_no_multimap/combined_count_df.rds"))
count_df <- combined_count_df |>
  dplyr::filter(sample_id_long == "CD34-WT-Cas9-dsODN-only-GSPneg-5uM-S90_S98") |>
  dplyr::rename("count" = "umi_count")
grna_info <- readRDS(paste0(data_dir, "spacer_info.rds"))
grna_spacer <- grna_info$spacer_seq
pam <- grna_info$pam
res <- find_guideseq_edit_sites(count_df, grna_spacer, pam)
}
