% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_detection.R
\name{find_guideseq_edit_sites}
\alias{find_guideseq_edit_sites}
\title{Find GUIDE-seq edit sites}
\usage{
find_guideseq_edit_sites(
  count_df,
  window_size = 1000,
  p_val_calculation_method = "exact",
  multiplicity_adjustment = "BH",
  multiplicity_adjustment_threshold = 0.1,
  c_tukey_beta = 10,
  c_tukey_sigma = 10,
  chrs_to_keep = seq(1L, 22L)
)
}
\arguments{
\item{chrs_to_keep}{}
}
\description{
TO DO:
\enumerate{
\item optimize c_tukey_beta and c_tukey_sigma for our setting
\item consider binning-free approaches
\item Multiple samples and covariates
}
}
\examples{
umi_tab_fp <- "/Users/timbarry/research_offsite/external/crispr-quant/guideseq/count_tables/293T-SpRY-Cas9-dsODN-only-GSPneg-S81_S89_L001_count_table.rds"
count_df <- readRDS(umi_tab_fp)
res_cntrl <- find_guideseq_edit_sites(count_df)

umi_tab_fp <- "/Users/timbarry/research_offsite/external/crispr-quant/guideseq/count_tables/293T-SpRY-Cas9-1620-GSPneg-S79_S87_L001_count_table.rds"
count_df <- readRDS(umi_tab_fp)
res_trt <- find_guideseq_edit_sites(count_df)

}
