% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nb_outlier_detection.R
\name{find_guideseq_edit_sites}
\alias{find_guideseq_edit_sites}
\title{Find GUIDE-seq edit sites}
\usage{
find_guideseq_edit_sites(
  count_df,
  bin_genome = TRUE,
  window_size = 1000,
  max_chain_link = 50,
  multiplicity_adjustment = "BH",
  model = "nb",
  robust_mle = TRUE,
  multiplicity_alpha = 0.1,
  c_tukey_beta = 10,
  c_tukey_sigma = 10,
  chrs_to_keep = paste0("chr", seq(1L, 22L)),
  plot_col = c("dodgerblue3", "firebrick")[1]
)
}
\arguments{
\item{chrs_to_keep}{}
}
\description{
TO DO:
\enumerate{
\item optimize c_tukey_beta and c_tukey_sigma for our setting
\item consider binning-free approaches
\item Multiple samples and covariates
}
}
\examples{
data_dir <- .get_config_path("LOCAL_CRISPR_DE_DATA_DIR")

umi_tab_fp <- paste0(data_dir, "/guideseq/count_tables/293T-SpRY-Cas9-dsODN-only-GSPneg-S81_S89_L001_count_table.rds")
count_df <- readRDS(umi_tab_fp)
res_cntrl <- find_guideseq_edit_sites(count_df, bin_genome = FALSE, plot_col = "firebrick")

umi_tab_fp <- paste0(data_dir, "/guideseq/count_tables/293T-SpRY-Cas9-1620-GSPneg-S79_S87_L001_count_table.rds")
count_df <- readRDS(umi_tab_fp)
res_trt <- find_guideseq_edit_sites(count_df, bin_genome = FALSE)
}
