% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_outlier_detection.R
\name{find_guideseq_edit_sites_dm_sliding_window}
\alias{find_guideseq_edit_sites_dm_sliding_window}
\title{Find GUIDE-seq edit sites DM}
\usage{
find_guideseq_edit_sites_dm_sliding_window(
  count_df,
  window_size = 25,
  multiplicity_adjustment = "BH",
  multiplicity_alpha = 0.1,
  chrs_to_keep = seq(1L, 22L),
  fit = NULL,
  plot_col = c("dodgerblue3", "firebrick")[1]
)
}
\arguments{
\item{plot_col}{}
}
\description{
Find GUIDE-seq edit sites DM
}
\examples{
data_dir <- .get_config_path("LOCAL_CRISPR_DE_DATA_DIR")

umi_tab_fp <- paste0(data_dir, "guideseq/count_tables/293T-SpRY-Cas9-1620-GSPneg-S79_S87_L001_count_table.rds")
count_df <- readRDS(umi_tab_fp)
res <- find_guideseq_edit_sites_dm_sliding_window(count_df)

umi_tab_fp <- paste0(data_dir, "guideseq/count_tables/293T-SpRY-Cas9-1620-GSPneg-S79_S87_L001_count_table.rds")
count_df <- readRDS(umi_tab_fp)
res_trt <- find_guideseq_edit_sites_dm_sliding_window(count_df)
}
