% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplicon_seq.R
\name{generate_synthetic_amplicon_seq_data}
\alias{generate_synthetic_amplicon_seq_data}
\title{Generate synthetic amplicon-seq data}
\usage{
generate_synthetic_amplicon_seq_data(
  p,
  r,
  pi_cntrl,
  editing_rate,
  n_amplicons_nonzero_editing,
  beta_binom_rho,
  sample_size_mu = 100000L,
  sample_size_theta = 15L
)
}
\arguments{
\item{p}{number of amplicons}

\item{r}{number of treatment/control samples}

\item{pi_cntrl}{(\link{0,1} scalar) background editing rate to use across all control samples}

\item{editing_rate}{(\link{0,1} scalar) editing rate in samples where there is nonzero editing}

\item{n_amplicons_nonzero_editing}{(integer) number of amplicons}

\item{beta_binom_rho}{(positive scalar) dispersion parameter for the beta-binomial distribution}

\item{sample_size_mu}{(positive scalar) mean sample size}

\item{sample_size_theta}{(positive scalar) size (i.e., overdispersion) parameter for sample size}
}
\description{
Generate synthetic amplicon-seq data
}
