% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_data.R
\name{generate_synthetic_multivariate_amplicon_seq_data}
\alias{generate_synthetic_multivariate_amplicon_seq_data}
\title{Generate synthetic multivariate amplicon-seq data}
\usage{
generate_synthetic_multivariate_amplicon_seq_data(
  n_mutation_types,
  n_alleles,
  n_covariates,
  pi_block,
  theta,
  phi_block,
  gamma_mat,
  delta_mat,
  covariate_bound,
  rho,
  r,
  mu_reads,
  theta_reads
)
}
\arguments{
\item{delta}{}
}
\description{
Generate synthetic multivariate amplicon-seq data
}
\examples{
n_mutation_types <- 2L
n_alleles <- c(50L, 55L)
pi_block <- c(0.98, 0.01, 0.01)
theta <- 0.75
phi_block <- c(0.5, 0.5)
gamma_mat <- matrix(c(-1.0, 0.1, -0.8, 0.2), nrow = 2L, byrow = TRUE)
delta_mat <- matrix(c(-3, 0.05, -2.75, 0.15), nrow = 2L, byrow = TRUE)
rownames(gamma_mat) <- rownames(delta_mat) <- c("mutation_type_1", "mutation_type_2")
colnames(gamma_mat) <- colnames(delta_mat) <- c("covariate_1", "covariate_2")
covariate_bound <- c(-1, 1)
rho <- 0.001
n_covariates <- 2L
r <- 6L
mu_reads <- 50000L
theta_reads <- 10
sim_data <- generate_synthetic_multivariate_amplicon_seq_data(n_mutation_types = n_mutation_types,
n_alleles = n_alleles, n_covariates = n_covariates, pi_block = pi_block,
theta = theta, phi_block = phi_block, gamma_mat = gamma_mat, delta_mat = delta_mat,
covariate_bound = covariate_bound, rho = rho, r = r, mu_reads = mu_reads,
theta_reads = theta_reads)
}
