% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_pt_3.R
\name{make_guideseq_qq_plot}
\alias{make_guideseq_qq_plot}
\title{Make GUIDE-seq QQ plot}
\usage{
make_guideseq_qq_plot(
  res_df,
  color_ground_truth = FALSE,
  rev_log_trans = TRUE,
  min_p = 1e-16
)
}
\arguments{
\item{res_df}{a data frame with column \code{p_value} giving the p-value of a given locus}
}
\value{
a qq plot of the results
}
\description{
Make GUIDE-seq QQ plot
}
\examples{
set.seed(7)
# NULL DATA
pi <- c(0.05, 0.1, 0.02)
mu_vect <- c(10, 6, 15)
theta_vect <- c(2, 5, 0.3)
m <- 10000
null_dat <- simulate_multirep_guideseq_data(pi, mu_vect, theta_vect, m)

# ALTERNATIVE DATA
pi <- c(0.5, 0.8, 0.6)
mu_vect <- c(250, 200, 50)
theta_vect <- c(50, 50, 60)
m_alt <- 15
alt_dat <- simulate_multirep_guideseq_data(pi, mu_vect, theta_vect, m_alt)

# COMBINED DATA
Y_mat <- cbind(alt_dat, null_dat)
colnames(Y_mat) <- paste0("window_", seq_len(ncol(Y_mat)))
incorporate_occupancy_info <- TRUE
multiplicity_alpha <- 0.2

# RUN METHOD
res_df <- run_multireplicate_guideseq_method(Y_mat, incorporate_occupancy_info = TRUE, multiplicity_alpha = 0.2)
res_df$true_editing <- c(rep(TRUE, m_alt), rep(FALSE, m))
p_all <- make_guideseq_qq_plot(res_df, color_ground_truth = TRUE)
p_null <- make_guideseq_qq_plot(res_df[seq(m_alt+1, nrow(res_df)),], color_ground_truth = FALSE)
}
