% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_pt_3.R
\name{make_local_scatterplot}
\alias{make_local_scatterplot}
\title{Make local scatterplot}
\usage{
make_local_scatterplot(annotated_df_sub)
}
\arguments{
\item{annotated_df_sub}{the annotated data frame for one window. gRNA and DNA sequence information must be present.}
}
\value{
a ggplot of the local UMI count distribution
}
\description{
Make local scatterplot
}
\examples{
caliper_res <- readRDS("/Users/timbarry/research_offsite/projects/crisprde-project/guideseq/hyperparam_res_list.rds")
res_df <- caliper_res$elane_cd34_wtcas9_e3sa$tuning_res$selected_trt_run$res_df
window_id <- res_df |> dplyr::slice(1L) |> dplyr::pull(window)
annotated_df_sub <- caliper_res$elane_cd34_wtcas9_e3sa$annotated_clustered_count_df_trt |> dplyr::filter(window == window_id)
p <- make_local_scatterplot(annotated_df_sub)
}
