% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_pt_2.R
\name{make_local_scatterplot}
\alias{make_local_scatterplot}
\title{Make local scatterplot}
\usage{
make_local_scatterplot(
  count_df_sub,
  x_range = NULL,
  log_trans = FALSE,
  point_size = 1,
  col = c("dodgerblue3", "firebrick")[1],
  title = NULL,
  target_seq = NULL,
  grna_spacer = NULL,
  pam_strand = NULL,
  protospacer_range = NULL,
  pam_range = NULL,
  cut_base_range = NULL
)
}
\description{
Make local scatterplot
}
