% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicate_aware_guide_seq.R
\name{overlap_homology_and_result_dfs}
\alias{overlap_homology_and_result_dfs}
\title{Overlap homology and result df}
\usage{
overlap_homology_and_result_dfs(result_df, homology_df)
}
\arguments{
\item{result_df}{output of run_multireplicate_guideseq_method}

\item{homology_df}{a homology data frame (for instance, output of load_crispritz_output)}
}
\value{
an augmented version of \code{result_df} with homology information appended
}
\description{
Add colummns:
\itemize{
\item \code{has_homology_hit} (indicating whether there is a homology hit inside the window)
\item \code{homology_n_mismatches} (indicating the number of mismatches between aligned spacer and protospacer sequence)
\item \code{homology_n_bulges} (indicating number of bulges between aligned spacer and protospacer sequence)
\item \code{homology_posit} (indicating the position of the start of the protospacer)
\item \code{homology_strand} (indicating whether the protospacer is on the plus or minus strand)
\item \code{homology_dna} (aligned protospacer sequence)
\item \code{homology_gRNA} (aligned spacer sequence)
}
}
\examples{
homology_df <- load_crispritz_output("/Users/timbarry/research_offsite/external/bauer-lab/guideseq_elane/crispritz/crispritz_CCCCGGCAGAAACGTCCGCG.hg38.targets.txt")
elane_dir <- paste0(.get_config_path("LOCAL_BAUER_LAB_DATA_DIR"), "guideseq_elane/")
count_df_all <- readRDS(paste0(elane_dir, "count_tables_no_multimap/combined_count_df.rds")) |>
dplyr::filter(cell_type == "CD34" & cas9_variant == "wt_cas9" & replicate_id \%in\% 1:2, chr != "chrM")
clustered_count_df <- count_df_all |> dplyr::filter(treated) |> cluster_loci()
res_df <- clustered_count_df |> construct_replicate_count_table() |> run_multireplicate_guideseq_method()
}
