% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_crispresso_output.R
\name{process_crispresso_allele_table_cas9}
\alias{process_crispresso_allele_table_cas9}
\title{Process CRISPRESSO allele table}
\usage{
process_crispresso_allele_table_cas9(allele_table_fp)
}
\arguments{
\item{allele_table_fp}{the file path to the allele count table}
}
\value{
a data frame with columns \code{allele_id} (giving the aligned sequence concatenated to the reference sequence, mutation_type (insertion, deletion, or complex), and mutation length)
}
\description{
This function assumes that CRISPRESSO has been run with stringent settings, in particular with \code{--ignore_substitutions}, \verb{quantification_window_size 1}, \code{--quantification_window_center -3}
}
\examples{
allele_table_fp <- "/Users/timbarry/research_offsite/external/bauer-lab/rhampseq_bcl11a/crispresso_output_1617/GUIDE0724-africa-edited/CRISPResso_on_1617_OT_0000/Alleles_frequency_table_around_sgRNA_CTAACAGTTGCTTTTATCAC.txt"
allele_feature_table <- process_crispresso_allele_table_cas9(allele_table_fp)
}
