% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplicon_seq.R
\name{run_amplicon_seq_analysis}
\alias{run_amplicon_seq_analysis}
\title{Run amplicon-seq analysis}
\usage{
run_amplicon_seq_analysis(
  data_list,
  editing_threshold = 0.01,
  nominal_ci_coverage = 0.95,
  nominal_fdr = 0.1,
  rho = NULL,
  tail = "right"
)
}
\arguments{
\item{data_list}{a list containing n_mat_trt, n_mat_cntrl, k_mat_trt, k_mat_cntrl}

\item{editing_threshold}{(scalar \link{0,1}) test whether editing is greater than this threshold}
}
\value{
a data frame containing
}
\description{
Run amplicon-seq analysis
}
\examples{
data_list <- generate_synthetic_amplicon_seq_data(p = 20L, r = 3L, pi_cntrl = 0.05, editing_rate = 0.1,
                                                  n_amplicons_nonzero_editing = 2L, beta_binom_rho = 0.01)
result_df <- run_amplicon_seq_analysis(data_list) |>
  dplyr::mutate(amplicon_id = factor(amplicon_id, labels = seq_len(p), levels = seq_len(p)))
create_amplicon_seq_ci_plot(result_df)
}
