% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplicon_seq.R
\name{run_amplicon_seq_analysis}
\alias{run_amplicon_seq_analysis}
\title{Run amplicon-seq analysis}
\usage{
run_amplicon_seq_analysis(
  data_list,
  editing_threshold = 0.001,
  nominal_ci_coverage = 0.95,
  nominal_fdr = 0.1,
  global_rho = FALSE,
  rho = NULL,
  tail = "right",
  outlier_mad_thresh = 4,
  min_mutation_count = 10L
)
}
\arguments{
\item{data_list}{a list containing n_mat_trt, n_mat_cntrl, k_mat_trt, k_mat_cntrl, and amplicon_ids}

\item{editing_threshold}{(scalar \link{0,1}) test whether editing is greater than this threshold}
}
\value{
a data frame containing
}
\description{
Run amplicon-seq analysis
}
\examples{
set.seed(1)
p <- 25L
amplicon_ids <- factor(x = paste0("amplicon_", seq_len(p)), levels = paste0("amplicon_", seq_len(p)))
beta_binom_rho <- c(0.005, rep(5e-4, times = p - 1L))
data_list <- generate_synthetic_amplicon_seq_data(p = p, r = 3L, pi_cntrl = 0.05, editing_rate = 0.15,
                                                  n_amplicons_nonzero_editing = 2L, beta_binom_rho,
                                                  amplicon_ids = amplicon_ids)
res <- run_amplicon_seq_analysis(data_list)

# make plots
make_amplicon_seq_ci_plot(res$result_df) |> plot()
make_amplicon_seq_p_value_plot(res$result_df) |> plot()
make_pilot_dispersion_plot(res) |> plot()
}
