% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{run_read_qc_on_sample}
\alias{run_read_qc_on_sample}
\title{Run read QC on sample}
\usage{
run_read_qc_on_sample(sample_dir)
}
\arguments{
\item{sample_dir}{}
}
\description{
Run read QC on sample
}
\examples{
sample_dir <- paste0(.get_config_path("LOCAL_BAUER_LAB_DATA_DIR"),
"guideseq_elane/count_tables_no_multimap/CD34-WT-Cas9-ELANE-e3SA-GSPneg-5uM-S89_S97/")
}
