% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{run_read_qc_on_sample}
\alias{run_read_qc_on_sample}
\title{Run read QC on sample}
\usage{
run_read_qc_on_sample(sample_dirs)
}
\arguments{
\item{sample_dir}{}
}
\description{
Run read QC on sample
}
\examples{
sample_dirs <- paste0("/Users/timbarry/research_code/genethoff-nf/demo/results/",
c("Jing_AAVS1_n1-10_Donor-Seq_AAVS1_GSP_plus_1", "Jing_AAVS1_n1-10_Donor-Seq_IL2RG_GSP_minus_1"))
qc_df <- run_read_qc_on_sample(sample_dirs)

}
