% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{run_read_qc_on_sample}
\alias{run_read_qc_on_sample}
\title{Run read QC on sample}
\usage{
run_read_qc_on_sample(sample_dir)
}
\arguments{
\item{sample_dir}{}
}
\description{
Run read QC on sample
}
\examples{
sample_dir <- "/Users/timbarry/research_code/genethoff-nf/demo/results/Jing_AAVS1_n1-10_Donor-Seq_AAVS1_GSP_plus_1"
}
