% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicate_aware_guide_seq.R
\name{tune_hyperparameters}
\alias{tune_hyperparameters}
\title{Tune hyperparameters}
\usage{
tune_hyperparameters(
  Y_mat_trt,
  Y_mat_cntrl,
  c_grid = c(1, 5, 10, 25, 50, 100),
  lambda_grid = c(0, 5, 10, 20, 50, 100, 500),
  multiplicity_alpha = 0.2,
  max_false_discs = 1L
)
}
\arguments{
\item{max_false_discs}{}
}
\description{
Tune hyperparameters
}
\examples{
elane_dir <- paste0(.get_config_path("LOCAL_BAUER_LAB_DATA_DIR"), "guideseq_elane/")
count_tables_dir <- paste0(elane_dir, "count_tables/")
combined_count_df <- readRDS(paste0(elane_dir, "count_tables_no_multimap/combined_count_df.rds"))
condition_vector <- c("trt", "cntrl")
Y_mat_trt <- combined_count_df |>
 dplyr::filter(treated, cas9_variant == "hifi_cas9", chr != "chrM") |>
 construct_replicate_count_table()
Y_mat_cntrl <- combined_count_df |>
 dplyr::filter(!treated, cas9_variant == "hifi_cas9", chr != "chrM") |>
 construct_replicate_count_table()
out <- tune_hyperparameters(Y_mat_trt, Y_mat_cntrl)
}
