% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicate_aware_guide_seq.R
\name{tune_hyperparameters}
\alias{tune_hyperparameters}
\title{Tune hyperparameters}
\usage{
tune_hyperparameters(
  Y_mat_trt,
  Y_mat_cntrl,
  c_grid = c(1, 5, 10, 25, 50, 100),
  lambda_grid = c(0, 5, 10, 20, 50, 100, 500),
  multiplicity_alpha = 0.2,
  max_false_discs = 1L
)
}
\arguments{
\item{Y_mat_trt}{treated count matrix}

\item{Y_mat_cntrl}{control count matrix}

\item{c_grid}{robust hyperparm grid}

\item{lambda_grid}{lambda grid}

\item{multiplicity_alpha}{nominal fdr}

\item{max_false_discs}{maximum false discoveries permitted in the control condition}
}
\value{
a list with elements \code{selected_params}, \code{selected_trt_run}, \code{selected_cntrl_run}, \code{grid_results}
}
\description{
Tune hyperparameters
}
\examples{
elane_dir <- paste0(.get_config_path("LOCAL_BAUER_LAB_DATA_DIR"), "guideseq_elane/")
count_df_all <- readRDS(paste0(elane_dir, "count_tables_no_multimap/combined_count_df.rds")) |>
dplyr::filter(cell_type == "CD34" & cas9_variant == "wt_cas9" & replicate_id \%in\% 1:2, chr != "chrM")
Y_mat_trt <- count_df_all |> dplyr::filter(treated) |> cluster_loci() |> construct_replicate_count_table()
Y_mat_cntrl <- count_df_all |> dplyr::filter(treated) |> cluster_loci() |> construct_replicate_count_table()

hyperparam_out <- tune_hyperparameters(Y_mat_trt = Y_mat_trt, Y_mat_cntrl = Y_mat_cntrl, c_grid = c(5, 25), lambda_grid = c(5, 50))
}
